♻️ refactor(input): nexon input & response component more universal data directory move
